fixed compiler warnings and crash
fixed crash when deleting entities
changed all long long ints in loops to size_t
updated readme to show end key for ent removal
updated version number

+++ bobwars/Camera.cpp
@@ -39,12 +39,12 @@ void Camera::zoom(float step)
 	zoomLevel *= step;
 }
 
-void Camera::rotate(int degrees)
+void Camera::rotate(float degrees)
 {
 	view.rotate(degrees);
 }
 
-void Camera::setRotation(int degrees)
+void Camera::setRotation(float degrees)
 {
 	view.setRotation(degrees);
 }
